fix #277: add voted claims to claim list template
