+++ api/templates/claim_list.html
@@ -15,6 +15,20 @@
             </div>
         </div>
     {% endfor %}
+
+    <div class="row medium-8 large-7 columns">
+        <h3>You have voted on these claims:</h3>
+    </div>
+    {% for claim in voted_claims %}
+        <div class="row medium-8 large-7 columns">
+            <div class="callout primary">
+                <h3><a href="{{claim.issue.url}}">{{claim.issue.title}}</a></h3>
+                <a href="//{{current_site.domain}}{% url 'custom-urls:claim-status' pk=claim.id %}">
+                    Check on the status
+                </a>
+            </div>
+        </div>
+    {% endfor %}
 {% endblock %}
 
 {% block scripts %}
