+++ auctions/tests/model_tests.py
@@ -15,36 +15,10 @@ from model_mommy import mommy
 from ..models import Bid, Claim, Issue, Vote
 from ..models import Offer, OfferFee, Payout, PayoutFee
 
-
-class MarketTestCase(TestCase):
-    def setUp(self):
-        """
-        Set up the following market of Bids for a single bug:
-            url: http://github.com/codesy/codesy/issues/37
-            user1: ask 50,  offer 0
-            user2: ask 100, offer 10
-            user3: ask 0,   offer 30
-        """
-        self.url = 'http://github.com/codesy/codesy/issues/37'
-        self.issue = mommy.make(Issue, url=self.url)
-        self.user1 = mommy.make(settings.AUTH_USER_MODEL,
-                                email='user1@test.com')
-        self.user2 = mommy.make(settings.AUTH_USER_MODEL,
-                                email='user2@test.com')
-        self.user3 = mommy.make(settings.AUTH_USER_MODEL,
-                                email='user3@test.com')
-        self.bid1 = mommy.make(Bid, user=self.user1,
-                               ask=50, offer=0, url=self.url)
-        self.bid2 = mommy.make(Bid, user=self.user2,
-                               ask=100, offer=10, url=self.url)
-        self.bid3 = mommy.make(Bid, user=self.user3,
-                               ask=0, offer=30, url=self.url)
-
-    def _add_url(self, string):
-        return string.format(url=self.url)
+from . import MarketWithBidsTestCase, MarketWithClaimTestCase
 
 
-class IssueTest(MarketTestCase):
+class IssueTest(MarketWithBidsTestCase):
 
     def test_unicode(self):
         self.issue.state = 'unknown'
@@ -61,7 +35,7 @@ class SimpleBidTest(TestCase):
         self.assertEqual(len(offers), 0)
 
 
-class BidTest(MarketTestCase):
+class BidTest(MarketWithBidsTestCase):
     def test_ask_met(self):
         self.assertFalse(self.bid1.ask_met())
         mommy.make(Bid, ask=0, offer=10, url=self.url)
@@ -118,7 +92,7 @@ class BidTest(MarketTestCase):
         )
 
 
-class NotifyMatchersReceiverTest(MarketTestCase):
+class NotifyMatchersReceiverTest(MarketWithBidsTestCase):
 
     @fudge.patch('auctions.models.send_mail')
     def test_dont_email_self_when_offering_more_than_ask(self, mock_send_mail):
@@ -188,31 +162,7 @@ class NotifyMatchersReceiverTest(MarketTestCase):
         )
 
 
-class ClaimTest(TestCase):
-    def setUp(self):
-        """
-        Set up the following claim senarios
-            user1: asks 50
-            user2: offers 25
-            user3: offers 25
-        """
-        self.url = 'http://github.com/codesy/codesy/issues/37'
-        self.issue = mommy.make(Issue, url=self.url)
-
-        self.user1 = mommy.make(settings.AUTH_USER_MODEL,
-                                email='user1@test.com')
-        self.user2 = mommy.make(settings.AUTH_USER_MODEL,
-                                email='user2@test.com')
-        self.user3 = mommy.make(settings.AUTH_USER_MODEL,
-                                email='user3@test.com')
-        self.bid1 = mommy.make(Bid, user=self.user1,
-                               ask=50, offer=0, url=self.url)
-        self.bid2 = mommy.make(Bid, user=self.user2,
-                               ask=0, offer=25, url=self.url)
-        self.bid3 = mommy.make(Bid, user=self.user3,
-                               ask=0, offer=25, url=self.url)
-        self.claim = mommy.make(Claim, user=self.user1, issue=self.issue)
-
+class ClaimTest(MarketWithClaimTestCase):
     def test_default_values(self):
         test_claim = mommy.make(Claim)
         self.assertEqual('Submitted', test_claim.status)
@@ -278,10 +228,10 @@ class ClaimTest(TestCase):
         )
 
 
-class NotifyMatchingOfferersTest(MarketTestCase):
+class NotifyMatchingOfferersTest(MarketWithBidsTestCase):
     def setUp(self):
         """
-        Add a final bid and a user2 claim to the MarketTestCase,
+        Add a final bid and a user2 claim to the MarketWithBidsTestCase,
         so the final market for the bug is now:
             url: http://github.com/codesy/codesy/issues/37
             user1: ask 50,  offer 0  (ask is met)
